Prepare declaration model for multiple scopes
Many elements can accept a range of scopes, so a "this is a lexical
declaration" role doesn't really model things well. Instead, remove it
and just have RakuAST::Declaration carry a scope. Also, nodes can now
specify their default scope and their acceptable scopes; the latter will
be used for generalized error checking of mis-scoped things later on.
This prepares the way for implementing various other scopes; also, do a
basic implementation of `anon`-scoped routines (although the
installation of the symbol name inside of the routine for anonymous
recursion is not yet implemented, and awaits implicit declaration
support).