Completely mimic Raku::Grammar hashes in OperatorProperties
This encapsulates all data needed for parsing Raku source into the
OperatorProperties class.  This adds the attributes:
- dba
- next-term
- sub-precedence
- diffy
- fiddly

This renames BuiltinOperatorTypes.iffy to multiplicative-iffy, as it
is basically the same as multiplicative, but with :iffy.

This add BuiltinOperatorTypes methods:
- dotty-infix
- additive-iffy
- junctive-and-iffy
- junctive-or-iffy
- named-unary
- tight-or-defor
- conditional
- conditional-ff
- list-assignment
- loose-or
- loose-or-xor
- sequencer

And fixes various incorrectly categorized infix ops, following the
information from the hashes in Raku::Grammar.